chore: moves TrestleRule dataclass to a new file
Improve error handling
Fixes my linting error
